Recur in .rollback() rather than .created_tables()
The .rollback() calls are only propagated back to UploadTask, which then
has to follow recursive dependencies to find objects to drop and tables
to truncate. When .deps() was changed back to direct dependencies, this
stopped working. But Script.created_tables() should let us do things
like check that each view is created in at most one script, so it
shouldn't recur; instead, let's recur in .rollback().